set pug view engine and started adding vue
