+++ Dockerfile
@@ -11,5 +11,4 @@ RUN NODE_ENV=null npm install
 
 ADD . /app
 
-EXPOSE 1337
 CMD ["npm", "start"]
