+++ docker-compose.yml
@@ -2,6 +2,8 @@ version: '2'
 services:
   bsgame:
     build: .
+    volumes:
+      - .:/app
     environment:
       PORT: 1337
       DATABASE_URI: postgres://bsgame@postgres/bsgame_development
