+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "repository": "https://github.com/Scoutski/bsgame",
   "scripts": {
-    "start": "node ./index.js",
+    "start": "node ./src/index.js",
     "test": "npm run pretest && npm run runtest; npm run posttest",
     "pretest": "sequelize db:migrate --url postgres://bsgame:@localhost:5432/bsgame_testing",
     "runtest": "env DATABASE_URI=postgres://bsgame:@localhost:5432/bsgame_testing ava --verbose",
@@ -17,8 +17,10 @@
     "bcrypt": "^1.0.3",
     "express": "^4.16.2",
     "pg": "^7.4.0",
+    "pug": "^2.0.0-rc.4",
     "sequelize": "^4.26.0",
-    "uuid": "^3.1.0"
+    "uuid": "^3.1.0",
+    "vue": "^2.5.9"
   },
   "devDependencies": {
     "ava": "^0.24.0",
