+++ src/views/index.pug
@@ -1,0 +1,6 @@
+html
+head
+  title='Vue App!'
+body
+  #app
+  script(src="./app.js")
