+++ src/vue/app.js
@@ -1,0 +1,7 @@
+import Vue from 'vue';
+
+const vm = new Vue({
+  // options
+});
+
+export default vm;
