+++ src/vue/client-entry.js
@@ -1,0 +1,4 @@
+import { app } from './app';
+
+console.log('vue app file loaded');
+app.$mount('#app');
